requirements.txt
python-dotenv~=0.21.1
requests~=2.28.2
substrate-interface~=1.5.0